requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
djangorestframework==3.12.4
idna==2.10
osmnx==0.16.2
numpy==1.20.2
pandas==1.2.4
Django>=3.1.2
psycopg2>=2.8.6,<2.8.7
gunicorn==20.0.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
